% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copaIntE.R
\name{copaIntE}
\alias{copaIntE}
\title{copaIntE}
\usage{
copaIntE(expressionSet, tails, thres = 0.05, method='Tibshirani',
corr=FALSE, offsets=NULL)
}
\arguments{
\item{expressionSet}{object containing  Set of matrices of molecular data and
phenotype data (1 for case, 0 for control)}

\item{tails}{Vector equal to number of matrices with values left or right for
where to find outliers}

\item{thres}{alpha value}

\item{method}{Tibshirani , Rank}

\item{corr}{Whether to correct for normal outliers}

\item{offsets}{A vector equal to the number of matrices which sets the minimum
value relative to normal to call outlier (corrected rank only)}
}
\value{
A vector with outlier counts by gene
}
\description{
Counts outliers by Tibshirani-Hastie method by calling outCount after setting
up list or by rank outlier method by calling outRank
}
\examples{
data(ExampleData)

 library(Biobase)
# building the Annotated Data Frame
 phenoData <- AnnotatedDataFrame(
     data.frame(
        type = factor(x = pheno, labels = c("Control", "Case")),
         row.names = colnames(expr)
     )
 )
# build environment
 inputData <- list2env(list(exprs = expr, meth = meth, cnv = cnv))

# build expressionSet - other information can be added here
 expressionSet <- ExpressionSet(inputData, phenoData)

# set up values for expr-meth-cnv in that order
 tailLRL <- c('left', 'right', 'left')


tibLRL <- copaIntE(expressionSet, tails=tailLRL)
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153
}

