% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expressionSetPheno.R
\name{expressionSetPheno}
\alias{expressionSetPheno}
\title{expressionSetPheno}
\usage{
expressionSetPheno(expressionSet)
}
\arguments{
\item{expressionSet}{ExpressionSet object containing sets of data and
phenotype information}
}
\value{
A list with all specific outlier calls for each molecular type in each
case sample
}
\description{
Returns pheno info
}
\examples{
data(ExampleData)

 library(Biobase)
# building the Annotated Data Frame
 phenoData <- AnnotatedDataFrame(
     data.frame(
        type = factor(x = pheno, labels = c("Control", "Case")),
         row.names = colnames(expr)
     )
 )
# build environment
 inputData <- list2env(list(exprs = expr, meth = meth, cnv = cnv))

# build expressionSet - other information can be added here
 expressionSet <- ExpressionSet(inputData, phenoData)

# retrieve phenotype data
 phenotype <- expressionSetPheno(expressionSet)
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153
}
\keyword{internal}

