\name{OGSA-package}
\alias{OGSA-package}
\alias{OGSA}
\docType{package}
\title{
This package uses outlier statistics and gene set analysis to identify deregulated pathways.
}
\description{
The package applies three versions of outlier statistics across multiple molecular
data types to create a single estimate at the gene level for the number of outliers 
relative to normal controls.  These gene estimates are used in gene set analysis to 
determine deregulated pathways.  Visualization of outlier calls provide sample 
specific information on potential drivers of the gene set statistic.
}
\details{
\tabular{ll}{
Package: \tab OGSA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-01\cr
License: \tab Gnu Public License\cr
}

}
\author{
Michael Ochs

Maintainer: Michael Ochs <ochsm@tcnj.edu>
}
\references{
Ochs MF, Farrar JE, Considine M, Wei Y, Meschinchi S, Arceci RJ. Outlier analysis and top scoring pair for integrated data analysis and biomarker discovery.  IEEE/ACM Trans Comput Biol and Bioinfo. 2014: 11: 520-32.
}
%%%%%%%%%%%%%%
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~

%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%%%%%%%%%%%%%%
%\examples{
%% ~~ simple examples of the most important functions ~~
%}
