% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outCallRank.R
\name{outCallRank}
\alias{outCallRank}
\title{outCallRank}
\usage{
outCallRank (dataSet, phenotype, thres= 0.05, tail='right', corr=FALSE,
offsets=NULL, names=NULL)
}
\arguments{
\item{dataSet}{Set of matrices of molecular data}

\item{phenotype}{A vector of 0s and 1s of length nSample, where 1 = case,
0 = control}

\item{thres}{Alpha value}

\item{tail}{A vector equal to the number of matrices with values left or right
for where to find outliers}

\item{corr}{Whether to correct for normal outliers}

\item{offsets}{A vector equal to the number of matrices which sets the minimum
value relative to normal to call outlier (corrected rank only)}

\item{names}{A vector equal to the number of matrices to name molecular type
of data (e.g., CNV)}
}
\value{
A list with all specific outlier calls for each molecular type in each
case sample
}
\description{
Counts outliers by the Ghosh method and generates list objects with all
outliers noted
}
\examples{
data(ExampleData)

#set up dataSet
dataSet <- list(expr, meth,cnv)

# Set up Phenotype
phenotype <- pheno
names(phenotype) <- colnames(cnv)

# set up values for expr-meth-cnv in that order
tailLRL <- c('left', 'right', 'left')

outRankLRL <- outCallRank(dataSet, phenotype, names=c('Expr',
                             'Meth', 'CNV'), tail=tailLRL)
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153

D. Ghosh. (2010). Discrete Nonparametric Algorithms for Outlier
Detection with Genomic Data. J. Biopharmaceutical Statistics, 20(2), 193-208.
}

