% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outCallRankE.R
\name{outCallRankE}
\alias{outCallRankE}
\title{outCallRankE}
\usage{
outCallRankE (expressionSet, thres= 0.05, tail='right', corr=FALSE,
offsets=NULL, names=NULL)
}
\arguments{
\item{expressionSet}{object containing  Set of matrices of molecular data and
phenotype data (1 for case, 0 for control)}

\item{thres}{Alpha value}

\item{tail}{A vector equal to the number of matrices with values left or right
for where to find outliers}

\item{corr}{Whether to correct for normal outliers}

\item{offsets}{A vector equal to the number of matrices which sets the minimum
value relative to normal to call outlier (corrected rank only)}

\item{names}{A vector equal to the number of matrices to name molecular type
of data (e.g., CNV)}
}
\value{
A list with all specific outlier calls for each molecular type in each
case sample
}
\description{
Counts outliers by the Ghosh method and generates list objects with all
outliers noted
}
\examples{
data(ExampleData)

 library(Biobase)
# building the Annotated Data Frame
 phenoData <- AnnotatedDataFrame(
     data.frame(
        type = factor(x = pheno, labels = c("Control", "Case")),
         row.names = colnames(expr)
     )
 )
# build environment
 inputData <- list2env(list(exprs = expr, meth = meth, cnv = cnv))

# build expressionSet - other information can be added here
 expressionSet <- ExpressionSet(inputData, phenoData)

# set up values for for the tails in the order that they are exported,
# for example:
tailLRL <- c('left', 'right', 'left')

outRankLRL <- outCallRankE(expressionSet, names=c('Expr', 'Meth', 'CNV'),
                                                          tail=tailLRL)
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153

D. Ghosh. (2010). Discrete Nonparametric Algorithms for Outlier
Detection with Genomic Data. J. Biopharmaceutical Statistics, 20(2), 193-208.
}

