% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outCallTib.R
\name{outCallTib}
\alias{outCallTib}
\title{outCallTib}
\usage{
outCallTib (dataSet, phenotype, tail='right', corr=FALSE, names=NULL)
}
\arguments{
\item{dataSet}{Set of matrices of molecular data}

\item{phenotype}{A vector of 0s and 1s of length nSample, where 1 = case,
0 = control}

\item{tail}{Vector equal to number of matrices with values 'left' or 'right'
for where to find outliers}

\item{corr}{whether to correct for normal outliers ONLY for compatibility,
since method does not allow determining specific changes in cases, it will
just print message if corr = TRUE}

\item{names}{Vector equal to number of matrices to name molecular type of data
(e.g., 'CNV').}
}
\value{
A list with all specific outlier calls for each molecular type in each
case sample
}
\description{
Counts outliers by the Tibshirani and Hastie method and generates a list
object with all outliers noted
}
\examples{
data(ExampleData)
data('KEGG_BC_GS')

# Set up dataSet
dataSet <- list(expr, meth, cnv)

# Set up Phenotype
phenotype <- pheno
names(phenotype) <- colnames(cnv)

# set up values for expr-meth-cnv in that order
tailLRL <- c('left', 'right', 'left')

outTibLRL <- outCallTib(dataSet, phenotype, names=c('Expr', 'Meth', 'CNV'), tail=tailLRL)
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153

D. Ghosh. (2010). Discrete Nonparametric Algorithms for Outlier
Detection with Genomic Data. J. Biopharmaceutical Statistics, 20(2), 193-208.
}

