% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OutCallTibE.R
\name{outCallTibE}
\alias{outCallTibE}
\title{outCallTibE}
\usage{
outCallTibE (expressionSet, tail='right', corr=FALSE, names=NULL)
}
\arguments{
\item{expressionSet}{ExpressionSet object containing sets of data and
phenotype information}

\item{tail}{Vector equal to number of matrices with values 'left' or 'right'
for where to find outliers}

\item{corr}{whether to correct for normal outliers ONLY for compatibility,
since method does not allow determining specific changes in cases, it will
just print message if corr = TRUE}

\item{names}{Vector equal to number of matrices to name molecular type of data
(e.g., 'CNV').}
}
\value{
A list with all specific outlier calls for each molecular type in each
case sample
}
\description{
Counts outliers by the Tibshirani and Hastie method and generates a list
object with all outliers noted
}
\examples{
data(ExampleData)
data('KEGG_BC_GS')

 library(Biobase)
# building the Annotated Data Frame
 phenoData <- AnnotatedDataFrame(
     data.frame(
        type = factor(x = pheno, labels = c("Control", "Case")),
         row.names = colnames(expr)
     )
 )
# build environment
 inputData <- list2env(list(exprs = expr, meth = meth, cnv = cnv))

# build expressionSet - other information can be added here
 expressionSet <- ExpressionSet(inputData, phenoData)

# set up values for for the tails in the order that they are exported, for example:
tailLRL <- c('left', 'right', 'left')


outTibLRL <- outCallTibE(expressionSet, names=c('Expr', 'Meth', 'CNV'), tail=tailLRL)
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153

D. Ghosh. (2010). Discrete Nonparametric Algorithms for Outlier
Detection with Genomic Data. J. Biopharmaceutical Statistics, 20(2), 193-208.
}

