% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outMap.R
\name{outMap}
\alias{outMap}
\title{outMap}
\usage{
outMap (outList, geneList, hmName = 'PatSpecMap.pdf', plotName =
'Outliers', truncGene = FALSE, clust=FALSE)
}
\arguments{
\item{outList}{List with all outliers generated by outCallRank or outCallTib}

\item{geneList}{Gene set to compare against}

\item{hmName}{Name for PDF output file}

\item{plotName}{Header for plot}

\item{truncGene}{if TRUE, only include genes that have outlier in the plot,
default is all genes in gene set}

\item{clust}{If TRUE, clusters data and produces dendrograms}
}
\value{
A matrix used for generating heatmap
}
\description{
Creates PDF color map of where outliers occur coded for molecular type
}
\examples{
data(ExampleData)
data('KEGG_BC_GS')

# Set up Phenotype
phenotype <- pheno
names(phenotype) <- colnames(cnv)

#set up datalist
dataSet <- list(expr,meth,cnv)

# set up values for expr-meth-cnv in that order
tailLRL <- c('left', 'right', 'left')

outTibLRL <- outCallTib(dataSet, phenotype=pheno,
                         names=c('Expr', 'Meth', 'CNV'), tail=tailLRL)

# put in your pathways here
pdgfB <- pathGS$'BIOCARTA_PDGF_PATHWAY'
outMap(outTibLRL, pdgfB, hmName='BC_PDGF_TIB.pdf', plotName='PDGF
Outlier T-H LRL Calls')
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153
}

