% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outRank.R
\name{outRank}
\alias{outRank}
\title{outRank}
\usage{
outRank (dataSet, phenotype, thres= 0.05, tail='right', corr=FALSE,
offsets=NULL)
}
\arguments{
\item{dataSet}{Set of matrices of molecular data}

\item{phenotype}{Vector of 1 for case, 0 for control}

\item{thres}{Alpha value}

\item{tail}{Vector equal to number of matrices with values 'left' or 'right'
for where to find outliers}

\item{corr}{Whether to correct for normal outliers}

\item{offsets}{Vector equal to number of matrices which sets minimum value
relative to normal to call outlier (corrected rank only)}
}
\value{
A vector with outlier counts by gene
}
\description{
Counts outliers by the Ghosh method.
}
\examples{
data(ExampleData)

# Set up Phenotype
phenotype <- pheno
names(phenotype) <- colnames(cnv)

#set up dataSet
dataSet <- list(expr, meth, cnv)

# set up values for expr-meth-cnv in that order
tailLRL <- c('left', 'right', 'left')

outRankLRL <- outRank(dataSet, phenotype, thres= 0.05, tail=tailLRL,
                             corr=FALSE, offsets=NULL)
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153

D. Ghosh. (2010). Discrete Nonparametric Algorithms for Outlier
Detection with Genomic Data. J. Biopharmaceutical Statistics, 20(2), 193-208.
}

