\name{pathGS}
\alias{pathGS}
\docType{data}
\title{
Gene set defined by BioCarta pathway
}
\description{
A large list containing gene set from the BioCarta pathway
}
\usage{pathGS}
\format{
List of 403 elements
}
\details{
Contained in KEGG_BC_GS data frame
}
\source{
Kanehisa, M. (2002). The KEGG databases at GenomeNet. Nucleic Acids Research, 30(1), 42-46. doi:10.1093/nar/30.1.42
}