% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testGScogps.R
\name{testGScogps}
\alias{testGScogps}
\title{testGScogps}
\usage{
testGScogps (outlierCts, geneSets)
}
\arguments{
\item{outlierCts}{Vector with gene names and outlier counts}

\item{geneSets}{List of gene sets}
}
\value{
A vector with rank sum gene set statistics
}
\description{
Performs gene set test on outlier counts
}
\examples{
\dontrun{
data(ExampleData)
data('_BC_GS')

#Set up your phenotype
phenotype <- rep(0, 69)
phenotype[annot[, 3] == 'Event'] <- 1
names(phenotype) <- rownames(annot)

# set up values for expr-meth-cnv in that order
tailLRL <- c('left', 'right', 'left')

dataSet <- list(expr, meth, cnv)

tibLRLcorr <- copaInt(dataSet, phenotype, tails=tailLRL, corr=TRUE)
gsTibLRLcorr <- testGScogps(tibLRLcorr, pathGS)
}
}
\references{
Ochs, M. F., Farrar, J. E., Considine, M., Wei, Y., Meshinchi, S.,
& Arceci, R. J. (n.d.). Outlier Analysis and Top Scoring Pair for Integrated
Data Analysis and Biomarker Discovery. IEEE/ACM Transactions on Computational
 Biology and Bioinformatics, 1-1. doi:10.1109/tcbb.2013.153
}

